requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
db==0.1.1
db-sqlite3==0.0.1
et_xmlfile==2.0.0
fastapi==0.115.12
h11==0.14.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
numpy==2.2.3
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.46.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
XlsxWriter==3.2.2
